+++ R/utils.R
@@ -2,4 +2,5 @@
 #' @importFrom shiny runApp
 #' @importFrom jsonlite toJSON
 #' @importFrom shinyTree shinyTree renderEmptyTree shinyTree get_selected updateTree
+#' @importFrom fs is_dir dir_ls path_real
 NULL