+++ inst/popkinr/nonmem_toolbox/server/home.R
@@ -1,11 +1,11 @@
 observeEvent(input$select_run, {
-    run_browser()$initialize_ui(force = TRUE)
+    run_browser()$initialize_ui()
 
     showModal(modalDialog(
       title = "Select a run folder or archive",
       size = "l",
       popkinr::serverBrowserUI("run_browser"),
-      div(tags$em("* Legend: Bold: directories containing *.tar.gz, *.tgz or *.zip archive files; Red: directories containing NONMEM run data")),
+
       footer = list(modalButton("Close"),
                     actionButton("load_run", "Load selection")),
       easyClose = TRUE)