+++ NEWS.md
@@ -1,3 +1,8 @@
+# popkinr 1.3
+
+* New server browser module
+* Removed run evaluation procedures (bootstrap, jacknife, chain) in `nonmem_monitor`
+
 # popkinr 1.1
 
 * Added a `NEWS.md` file to track changes to the package.